+++ docs/CHANGELOG.md
@@ -15,6 +15,11 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Changed
+- All `NEAT` importers does not have `is_reference` parameter in `.to_rules()` method. This has been moved
+  to the `ExcelExporter` `__init__` method. This is because this is the only place where this parameter was used.
+
 ## [0.75.5] - 24-05-24
 ### Fixed
 - Potential of having duplicated spaces are now fixed